requirements.txt
streamlit>=1.31.0
openai>=1.0.0
sentence-transformers>=2.6.0
faiss-cpu>=1.7.4
pymupdf>=1.23.0
python-dotenv>=1.0.1
pandas>=2.1.0
langchain
langgraph
requests
beautifulsoup4
